Match-id-cb8f66fd6b48ae2cf70dfd5ea9d6174f4dbed049

+++ csrc/event_trace/mstx_hooks/mstx_manager.cpp
@@ -111,7 +111,7 @@ void MstxManager::ReportRegionsRegister(mstxDomainHandle_t domain, mstxMemRegion
     const mstxMemVirtualRangeDesc_t *rangeDescArray =
         reinterpret_cast<const mstxMemVirtualRangeDesc_t *>(desc->regionDescArray);
 
-    for (int i = 0; i < desc->regionCount; i++) {
+    for (size_t i = 0; i < desc->regionCount; i++) {
         TorchNpuRecord torchNpuRecord;
         int devId = rangeDescArray[i].deviceId;
         memUsageMp_[devId].dataType = 0;
@@ -136,7 +136,7 @@ void MstxManager::ReportRegionsUnregister(mstxDomainHandle_t domain, mstxMemRegi
         return;
     }
     std::lock_guard<std::mutex> guard(mutex_);
-    for (int i = 0; i < desc->refCount; i++) {
+    for (size_t i = 0; i < desc->refCount; i++) {
         if (!regionHandleMp_.count(desc->refArray[i].pointer)) {
             continue;
         }
